CarBazaar-Car-Marketplace-Site-With-Valuation-Of-Used-Cars-Using-ML/NewBazaar/requirements.txt
streamlit == 1.27.0
catboost == 1.2.2
streamlit-authenticator == 0.2.2
streamlit-extras == 0.3.4
streamlit-option-menu == 0.3.6
streamlit-carousel == 0.0.4
imgbbpy == 0.1.3
pymongo == 4.5.0
st_clickable_images == 0.0.3
pillow-avif-plugin == 1.4.1